return circle packing to index
