+++ src/pages/index.tsx
@@ -8,11 +8,23 @@ import breakPoints from "../utils/breakpoints";
 import styles from "./index.module.scss";
 import Label from "../components/Input/Label";
 import { useQuery } from "react-query";
-import { getModelCount } from "../apis/AggregatedData.api";
+import { getCancerHierarchy, getModelCount } from "../apis/AggregatedData.api";
+import { useRouter } from "next/router";
 import Loader from "../components/Loader/Loader";
 import SearchBar from "../components/SearchBar/SearchBar";
 import dynamic from "next/dynamic";
 
+const DynamicCirclePacking = dynamic(
+	import("../components/CirclePacking/CirclePacking"),
+	{
+		loading: () => (
+			<div style={{ height: "300px" }}>
+				<Loader />
+			</div>
+		),
+		ssr: false,
+	}
+);
 const DynamicDataCountCard = dynamic(
 	import("../components/DataCountCard/DataCountCard"),
 	{
@@ -28,9 +40,13 @@ const DynamicDataCountCard = dynamic(
 const Home: NextPage = () => {
 	const { windowWidth } = useWindowDimensions();
 	let bpLarge = breakPoints.large;
+	let cancerHierarchyQuery = useQuery("cancerHierarchy", () => {
+		return getCancerHierarchy();
+	});
 	let modelCount = useQuery("modelCount", () => {
 		return getModelCount();
 	});
+	const router = useRouter();
 
 	return (
 		<>
@@ -83,6 +99,44 @@ const Home: NextPage = () => {
 			<section>
 				<div className="container">
 					<div className="row align-center">
+						<ShowHide windowWidth={windowWidth || 0} showOver={bpLarge}>
+							<div
+								className={`col-12 col-md-10 col-lg-5 offset-md-1 offset-lg-0 ${styles.circlePacking_col}`}
+							>
+								{/* Graph */}
+								<div
+									style={{
+										backgroundColor: "#085154",
+										aspectRatio: "1",
+										borderRadius: "500%",
+									}}
+								>
+									{!cancerHierarchyQuery.isLoading &&
+									cancerHierarchyQuery.data ? (
+										<DynamicCirclePacking
+											data={cancerHierarchyQuery.data}
+											onCircleClick={(circleId, circleDepth) => {
+												const searchPrefix =
+													circleDepth === 1
+														? `?filters=cancer_system:`
+														: `?filters=search_terms:`;
+												const termSuffix = circleDepth === 1 ? "Cancer" : "";
+												const search = `${searchPrefix}${encodeURIComponent(
+													circleId + termSuffix
+												)}`;
+
+												router.push({
+													pathname: "/search",
+													search,
+												});
+											}}
+										/>
+									) : (
+										<Loader />
+									)}
+								</div>
+							</div>
+						</ShowHide>
 						<div className="col-12 col-lg-6 offset-lg-1">
 							<h2>
 								We can collect and display your data making you more
